+++ src/js/listener.js
@@ -15,6 +15,8 @@ class Listener {
         this.loop = false;
 
         this.initLoadedSettings();
+        this.initEvents();
+        this.update();
     }
 
     initLoadedSettings() {
@@ -25,11 +27,59 @@ class Listener {
         }, 0);
     }
 
+    initEvents() {
+        this.audio.addEventListener("ended", () => {
+            this.nextButton();
+        });
+
+        this.audio.addEventListener("error", (e) => {
+            this.app.error(`ERROR HK-304 => ${e.target.error}`);
+        });
+
+        ipcRenderer.on("song-control", (e, data) => {
+            switch (data) {
+                case "play": this.playButton(); break;
+                case "next": this.nextButton(); break;
+                case "previous": this.previousButton(); break;
+                default: break;
+            }
+        });
+
+        this.app.elements.footer.center.song.slider.addEventListener("input", (e) => {
+            if (this.currentSong == null) {
+                this.app.elements.footer.center.song.slider.value = 0;
+                return;
+            }
+            const percentage = e.target.value;
+            const seconds = percentage / 100 * this.audio.duration;
+            this.audio.currentTime = seconds;
+        });
+    }
+
     // CLEANUP
     deconstructor() {
         this.app.data.settings.random = this.random;
         this.app.data.settings.loop = this.loop;
         this.app.data.settings.volume = this.audio.volume;
+    }
+
+    // UPDATE
+    update() {
+        this.updateSongSlider();
+        this.updateSongCurrentTime();
+        setTimeout(() => this.update(), 1000 / 30);
+    }
+
+    updateSongSlider() {
+        if (this.currentSong == null) return;
+        const percentage = this.audio.currentTime / this.audio.duration * 100;
+        if (isNaN(percentage)) return;
+        this.app.elements.footer.center.song.slider.value = percentage;
+    }
+
+    updateSongCurrentTime() {
+        if (this.currentSong == null) return;
+        this.app.elements.footer.center.song.position.textContent = formatTime(parseInt(this.audio.currentTime));
     }
 
     // BUTTON
@@ -124,15 +174,23 @@ class Listener {
         this.audio.src = songPath;
         this.audio.setAttribute("song-id", sID);
         this.audio.play();
+
+        const footer = this.app.elements.footer;
+        footer.left.title.textContent = this.currentSong.name;
+        footer.left.artist.textContent = this.currentSong.artist;
+
+        this.audio.addEventListener("loadeddata", () => footer.center.song.duration.textContent = formatTime(parseInt(this.audio.duration)));
     }
 
     displayPlayButton() {
         this.app.elements.footer.center.buttons.play.classList.remove("hidden");
         this.app.elements.footer.center.buttons.pause.classList.add("hidden");
+        ipcRenderer.send("set-thumbnail-play-button", "play");
     }
 
     displayPauseButton() {
         this.app.elements.footer.center.buttons.play.classList.add("hidden");
         this.app.elements.footer.center.buttons.pause.classList.remove("hidden");
+        ipcRenderer.send("set-thumbnail-play-button", "pause");
     }
 };