+++ src/js/main.js
@@ -1,8 +1,6 @@
-const { ipcRenderer, shell } = require("electron");
+const { ipcRenderer } = require("electron");
 const fs = require("fs");
 const fsp = require("fs").promises;
 const path = require("path");
 
-window.addEventListener("load", () => {
-    new App();
-});
+window.addEventListener("load", () => new App());